refactor: refactor sync schedule components placement
